backend/requirements.production.txt
# URADI-360 Production Requirements
# Optimized for production deployment

# =============================================================================
# Core Framework
# =============================================================================
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.6.0
email-validator==2.2.0
python-multipart==0.0.12

# =============================================================================
# Database
# =============================================================================
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.14.0

# =============================================================================
# Authentication & Security
# =============================================================================
pyjwt==2.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
bleach==6.2.0

# =============================================================================
# Caching & Task Queue
# =============================================================================
redis==5.2.0
celery==5.4.0
celery-redbeat==2.2.0

# =============================================================================
# Rate Limiting
# =============================================================================
slowapi==0.1.9

# =============================================================================
# HTTP Clients (for external APIs)
# =============================================================================
httpx==0.27.2
aiohttp==3.11.0
aiofiles==24.1.0

# =============================================================================
# AI/ML Services
# =============================================================================
openai==1.55.0
scikit-learn==1.5.2
hdbscan==0.8.40
langdetect==1.0.9
transformers==4.46.0
torch==2.5.0
sentence-transformers==3.3.0

# =============================================================================
# Vector Database
# =============================================================================
qdrant-client==1.12.0

# =============================================================================
# Web Scraping & Content Parsing
# =============================================================================
beautifulsoup4==4.12.3
feedparser==6.0.11
lxml==5.3.0
newspaper3k==0.2.8
trafilatura==1.12.0

# =============================================================================
# Social Media APIs
# =============================================================================
tweepy==4.14.0
facebook-sdk==3.1.0
instaloader==4.13.0

# =============================================================================
# Storage & File Handling
# =============================================================================
minio==7.2.12
python-magic==0.4.27
Pillow==11.0.0

# =============================================================================
# SMS & Notifications
# =============================================================================
twilio==9.3.0

# =============================================================================
# Monitoring & Observability
# =============================================================================
sentry-sdk[fastapi]==2.19.0
structlog==24.4.0
psutil==6.1.0
prometheus-client==0.21.0

# =============================================================================
# Scheduling & Date Handling
# =============================================================================
python-dateutil==2.9.0
pytz==2024.2

# =============================================================================
# Data Processing
# =============================================================================
pandas==2.2.0
numpy==1.26.0

# =============================================================================
# Utilities
# =============================================================================
tenacity==9.0.0
validators==0.34.0

# =============================================================================
# Testing (only needed for CI/CD, can be removed for minimal image)
# =============================================================================
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2
